# Documentacion/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Vida Cotidiana API
  version: 1.0.0
  description: >
    Canonical API contract for Vida Cotidiana V1. All timestamps are
    ISO-8601 in UTC. All identifiers are UUIDv4. Authentication is delegated
    to Keycloak (OIDC/OAuth 2.1, Authorization Code + PKCE, DEC-004/ADR-008);
    this API is a pure OAuth2 resource server and defines no /auth/* login
    endpoint of its own. An operational healthcheck (e.g. Spring Boot
    Actuator /actuator/health) exists outside this versioned business
    contract and is intentionally not documented here (see
    07-backend-architecture.md).
servers:
  - url: /api/v1
security:
  - keycloakBearer: []
paths:
...
  /vision-boards:
    get:
      summary: List own vision boards (paginated). Items omit `elements` — see GET /vision-boards/{id}.
      security:
        - keycloakBearer: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated vision board list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    additionalProperties: true # BE-035, see PageMeta
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/VisionBoard'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a vision board (caller becomes owner)
      security:
        - keycloakBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisionBoardRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionBoard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /vision-boards/{id}:
    get:
      summary: Get a vision board (owner only), including its elements
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vision board with elements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionBoard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a vision board (owner only). Requires the current version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVisionBoardRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionBoard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking) — the vision board was modified concurrently; refetch and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a vision board (owner only). Cascades to its elements at the database level.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /vision-boards/{id}/elements:
    post:
      summary: Add an element to a vision board (owner only)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateElementRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionBoardElement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /vision-boards/{id}/elements/{elementId}:
    put:
      summary: Update an element (owner only, via the board). Requires the current version of the element for optimistic locking — independent of the board's own version.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: elementId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateElementRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionBoardElement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking) — the element was modified concurrently; refetch and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an element (owner only, via the board)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: elementId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /vision-boards/{id}/elements/{elementId}/reorder:
    post:
      summary: >
        Move an element exactly one position relative to its immediately
        adjacent neighbor in paint order (RAISE/LOWER), or to the very
        front/back (FRONT/BACK). Owner only, via the board. Atomically
        renumbers every element on the board to a dense zIndex sequence
        matching the new order and returns all of them, so the caller can
        sync every affected element without a full board refetch. Requires
        the target element's current version for optimistic locking — the
        renumbering of its siblings is a side effect of this one action, not
        something the caller is expected to have a version for.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: elementId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderElementRequest'
      responses:
        '200':
          description: The board's elements, all reflecting their (possibly new) zIndex.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisionBoardElement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking) — the element was modified concurrently; refetch and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    keycloakBearer:
      type: oauth2
      description: >
        OAuth2/OIDC access token issued by Keycloak (Authorization Code +
        PKCE from the client). The backend validates the bearer JWT on every
        request as a pure OAuth2 resource server; it never issues its own
        session or token (DEC-004/ADR-008). Token issuer/JWK Set URL are
        environment configuration, not part of this contract.
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.example/realms/vida-cotidiana/protocol/openid-connect/auth
          tokenUrl: https://keycloak.example/realms/vida-cotidiana/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect authentication
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index.
    SizeParam:
      in: query
      name: size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size (max 100).
  schemas:
    Error:
      type: object
      description: Uniform error envelope. Never includes stack traces, SQL, internal names or secrets.
      required: [code, message, traceId]
      properties:
        code:
          type: string
          example: REMINDER_NOT_FOUND
        message:
          type: string
          example: The requested reminder was not found.
        traceId:
          type: string
    PageMeta:
      type: object
      # BE-035: additionalProperties: true is required here because PageMeta is combined via allOf
      # with a sibling object (page items) at each paginated endpoint — without it, strict allOf
      # validators (incl. swagger-request-validator/TEST-API-001) reject each branch as a closed
      # schema and fail the sibling's fields as "not allowed", even though the merged instance is
      # exactly what the composed schema intends. Not a loosening of type/required validation.
      additionalProperties: true
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    CreateVisionBoardRequest:
      type: object
      required: [name, width, height]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        theme:
          type: string
          enum: [LIGHT, DARK, PAPER, NATURAL, CALM, ENERGY]
          description: >
            Board Theme (FASE 16) — independent of the app's own visual
            theme. Optional on create; defaults to LIGHT when omitted. The
            actual color palette per theme is defined once, centrally, on
            the frontend (visionBoardThemes.ts) — this is only ever the
            stable id, same contract as iconId/stickerId elsewhere in this API.
    UpdateVisionBoardRequest:
      type: object
      description: >
        Partial update para name/description/width/height/theme
        (opcionales, null deja el valor guardado sin cambios) — así también
        se persiste un cambio de Board Theme (FASE 16): el cliente envía
        solo { theme, version }, no existe un endpoint dedicado para esto.
        version es obligatorio (bloqueo optimista).
      required: [version]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        theme:
          type: string
          enum: [LIGHT, DARK, PAPER, NATURAL, CALM, ENERGY]
        version:
          type: integer
          description: Expected current version of the vision board; a mismatch returns 409.
    VisionBoard:
      type: object
      required: [id, name, width, height, theme, version]
      description: Owner-only entity — no sharing/collaborator concept, mismo criterio que Note/Warranty.
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        width:
          type: integer
        height:
          type: integer
        theme:
          type: string
          enum: [LIGHT, DARK, PAPER, NATURAL, CALM, ENERGY]
          description: Board Theme (FASE 16) — always present, defaults to LIGHT.
        version:
          type: integer
          description: Optimistic locking version, incremented on every update.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        elements:
          type: array
          description: Only populated by GET /vision-boards/{id} — omitted on the list endpoint (GET /vision-boards).
          items:
            $ref: '#/components/schemas/VisionBoardElement'
    CreateElementRequest:
      type: object
      required: [type, x, y, width, height]
      properties:
        type:
          type: string
          enum: [TEXT, IMAGE, NOTE, STICKER, SHAPE]
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        data:
          type: object
          additionalProperties: true
          description: Type-specific payload (font/color for TEXT, url/crop for IMAGE, etc.) — shape varies per `type`, validated client-side. Optional; defaults to an empty object.
    UpdateElementRequest:
      type: object
      description: >
        Partial update para x/y/width/height/rotation/zIndex/locked/visible
        (opcionales, null deja el valor guardado sin cambios). data NO es
        parcial — siempre se aplica tal cual se envía cuando está presente
        (mismo contrato que iconId/stickerId en otras entidades). version es
        obligatorio (bloqueo optimista, independiente del version del board).
      required: [version]
      properties:
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        rotation:
          type: number
          format: double
        zIndex:
          type: integer
        locked:
          type: boolean
        visible:
          type: boolean
        data:
          type: object
          additionalProperties: true
        version:
          type: integer
          description: Expected current version of the element; a mismatch returns 409.
    ReorderElementRequest:
      type: object
      description: >
        direction is FRONT (al frente), BACK (al fondo), RAISE (subir una
        capa) o LOWER (bajar una capa) — RAISE/LOWER swap the element with
        whichever one is immediately above/below it in paint order. version
        es obligatorio (bloqueo optimista sobre el elemento objetivo).
      required: [direction, version]
      properties:
        direction:
          type: string
          enum: [FRONT, BACK, RAISE, LOWER]
        version:
          type: integer
          description: Expected current version of the target element; a mismatch returns 409.
...
    VisionBoardElement:
      type: object
      required: [id, boardId, type, x, y, width, height, rotation, zIndex, locked, visible, version]
      properties:
        id:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        type:
          type: string
          enum: [TEXT, IMAGE, NOTE, STICKER, SHAPE]
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        rotation:
          type: number
          format: double
        zIndex:
          type: integer
        locked:
          type: boolean
        visible:
          type: boolean
        data:
          type: object
          additionalProperties: true
        version:
          type: integer
          description: Optimistic locking version, incremented on every update — independent of the board's own version.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Input validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated (missing or invalid bearer token).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource does not exist or is not accessible to the caller (never distinguishes the two, to avoid leaking existence, SEC-001).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...